backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
channels==4.3.0
daphne==4.2.0
psycopg[binary]==3.3.1
django-cors-headers==4.9.0
python-dotenv==1.2.0
Pillow>=10.0
whitenoise==6.9.0
drf-spectacular==0.29.0
